+++ etna/models/linear.py
@@ -72,6 +72,8 @@ class LinearPerSegmentModel(
         fit_intercept:
             Whether to calculate the intercept for this model. If set to False, no intercept will be used in
             calculations (i.e. data is expected to be centered).
+        **kwargs:
+            Additional parameters passed to :py:class:`sklearn.linear_model.LinearRegression` model.
         """
         self.fit_intercept = fit_intercept
         self.kwargs = kwargs
@@ -126,6 +128,8 @@ class ElasticPerSegmentModel(
         fit_intercept:
             Whether to calculate the intercept for this model. If set to False, no intercept will be used in
             calculations (i.e. data is expected to be centered).
+        **kwargs:
+            Additional parameters passed to :py:class:`sklearn.linear_model.ElasticNet` model.
         """
         self.alpha = alpha
         self.l1_ratio = l1_ratio
@@ -175,6 +179,8 @@ class LinearMultiSegmentModel(
         fit_intercept:
             Whether to calculate the intercept for this model. If set to False, no intercept will be used in
             calculations (i.e. data is expected to be centered).
+        **kwargs:
+            Additional parameters passed to :py:class:`sklearn.linear_model.LinearRegression` model.
         """
         self.fit_intercept = fit_intercept
         self.kwargs = kwargs
@@ -229,6 +235,8 @@ class ElasticMultiSegmentModel(
         fit_intercept:
             Whether to calculate the intercept for this model. If set to False, no intercept will be used in
             calculations (i.e. data is expected to be centered).
+        **kwargs:
+            Additional parameters passed to :py:class:`sklearn.linear_model.ElasticNet` model.
         """
         self.alpha = alpha
         self.l1_ratio = l1_ratio